120. 정범호 / 애니메이션 모델 Space Rotation 값을 계산하여 리턴해주는 함수, 로직추가, Slient몹 상태클래스만추가, Sprise몹 문열릴때 내려오는 로직추가.

+++ TWEngine/Include/Component/ColliderOBB3D.cpp
@@ -195,7 +195,7 @@ void CColliderOBB3D::Save(BinaryWrite * _pInstBW)
 	_pInstBW->WriteData(m_strCollisionGroup);
 	_pInstBW->WriteData(m_MyTypeName);
 	
-	int iVecExceptTypeNameSize = m_vecContinueName.size();
+	int iVecExceptTypeNameSize = (int)m_vecContinueName.size();
 	_pInstBW->WriteData(iVecExceptTypeNameSize);
 
 	for (size_t i = 0; i < iVecExceptTypeNameSize; ++i)
